requirements.txt
asyncio==3.4.3
discord.py==0.16.12
requests==2.13.0
